moving out the cover logic the a separate file
the goal is to have a cleaner component and don't re-read the whole
directory every time a cover is rendered.

A new feature is to throw an error if a file is found that cannot
be paired with any book.


this implementation a little cursed with lower case strings.
will clean that up