add a basic tag page
this is a copy paste of the table component with small changes.
I should probably refactor the common parts in the future but
that is more work and I am not even sure at this point that
a page like this is going to be useful so I will keep playing
with it and see.